Added the function "usageScenarios":
Summarize losses (sw take + gw sd depletion) per scenario for each
lowflow site using the WAPs within the associated catchment of that
lowflow site (without upstream lowflow site catchments).

Uses a csv-file with the header corresponding to the scenarios, and the
first column indicating the lowflow sites. The WAPs that need to be
summarized are then found as values in the column. The last to columns
of the input csv-file are consent numbers (if waps are shared among
multiple consents) and block (A or B).

Output is a csv-file for each scenario (column in input file) with the
summed surface water loss for the WAPs found in the catchment belonging
to the lowflow site in each column.